Initial Commit
As of this commit, Inca Tool (it) is completely functional. It has
been tested (with a lot of switches) and proven to be solid.
Future plans include a task file to replace all the long flags
that need to be used, and not require a raw source file for a task
but instead use a list of commands and compile them for the user.