Issue #3 - Added includes to both task files and inventory list
The syntax is `@/path/to/file` with @ being the first character
on the line. For both file types, an include acts as if the
text from the included file replaced the include line. The path
is relative to the working directory, not the task or inventory file.
If a file tries to include itself, it will fail.

The device list is implemented such that the full text data is loaded
into memory and then processed. For task files, the file is processed
and includes are processed inline. The way the parsers were created
lended themselves to either solution. It would be nice one day to
rewrite one of the parsers (probably the inventory list) and make
it in line with the other. That way the text doesn't need to be loaded
into memory.