Closes #3 - Add executable includes to device list
Using the syntax `@!/path/to/executable` will execute the file
and add its standard output to the inventory file. This can be
used to dynamically generate a list or groups. If the file doesn't
exist or an error occures, the parse will fail.

Also, actually fail if an include file doesn't exist or fails itself.