Renamed packages and created a more logical package structure
The script generation was being split between two packages when
it's one job. The entire script generation process is now in the
script package. The task package only deals with parsing a task file
and housing the Task struct. taskmanager is renamed to just manager.
The _s command syntax works as documented, the script is ran once
not for every device.